+++ src/router/index.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from 'react-router-dom'
 
 import LoginPage from '@/components/auth/Login'
 import SignupPage from '@/components/auth/Signup'
-import CardDetail from '@/components/card/CardDetail'
+
 import CategoryContainer from '@/components/category/CategoryContainer'
 import HomePage from '@/components/home/Home'
 import NotificationPage from '@/components/inbox/NotificationPage'
@@ -10,6 +10,7 @@ import ProjectMainContainer from '@/components/projectMain/ProjectMainContainer'
 import ProjectUpdateContainer from '@/components/projectUpdate/ProjectUpdateContainer'
 import SectionContainer from '@/components/section/SectionContainer'
 import { AuthLayout, LandingLayout, MainLayout } from '@/layout/index'
+import CardDetailContainer from '@/components/card/CardDetailContainer'
 
 const router = createBrowserRouter([
   {
@@ -73,11 +74,11 @@ const router = createBrowserRouter([
                 children: [
                   {
                     index: true,
-                    element: <CardDetail />,
+                    element: <CardDetailContainer />,
                   },
                   {
                     path: 'edit',
-                    element: <CardDetail mode="edit" />,
+                    element: <CardDetailContainer mode="edit" />,
                   },
                 ],
               },
